[services] Tunnelbroker - Refactor time types
Summary:
We want to change `long long` to `int64_t` to be sure about the size of variables(long basically might be compiler-dependent).

https://linear.app/comm/issue/ENG-584/refactor-time-types-in-c

Test Plan:
```
cd services
yarn run-tunnelbroker-service
```

Reviewers: geekbrother, palys-swm

Reviewed By: geekbrother, palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3711

+++ services/tunnelbroker/src/Amqp/AmqpManager.h
@@ -15,7 +15,7 @@ class AmqpManager {
 
   std::unique_ptr<AMQP::TcpChannel> amqpChannel;
   std::atomic<bool> amqpReady;
-  std::atomic<long long> lastConnectionTimestamp;
+  std::atomic<int64_t> lastConnectionTimestamp;
   void connectInternal();
 
 public:
